🌱 Filter PRs by dates instead of milestones
Since we no longer use milestones on PRs, I had to find another way to
get PRs of the last sprint.

The solution is to filter PRs with mergedAt in between sprintEnd and
sprintEnd - sprintLengthDays
where sprintEnd and sprintLengthDays are optional command line arguments
with defaults sprintEnd = today; sprintLengthDays = 12
since our sprints start on Monday and end on Friday

Since the second parameter (milestoneId) is not needed anymore, I
changed it to projectName, so we can run the report also for other
projects, e.g. lisk-service